requirements.txt
absl-py==2.1.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
astroid==3.1.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
dm-tree==0.1.8
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.13.1
flake8==7.0.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==24.3.7
fonttools==4.50.0
fqdn==1.5.1
fsspec==2024.3.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.10.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.3.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.0.5
kiwisolver==1.4.5
lark==1.2.2
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
ml-dtypes==0.3.2
mlxtend==0.23.4
mpmath==1.3.0
namex==0.0.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.5
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
protobuf==4.25.3
psutil==7.1.0
pure_eval==0.2.3
pycodestyle==2.11.1
pycparser==2.23
pyflakes==3.2.0
Pygments==2.17.2
pylint==3.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2024.1
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.7.1
rpds-py==0.27.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.3.0
tinycss2==1.4.0
tomlkit==0.12.4
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.43.0
wrapt==1.16.0
